Removes getDelegatorSubscribers and getListOfDelegateAddressAndDelegatorAddress from Subscriber.model to avoid circular dependency
